+++ views/layouts/main.hbs
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>The Greatest Store</title>
-  <link rel="stylesheet" href="/public/css/styles.css">
+  <link rel="stylesheet" type="text/css" href="/public/css/styles.css">
 </head>
 
 <body>
